List of functions in BaroclinicAdjustmen

BaroclinicAdjustment.Parameterizations.BiharmonicLeith — Type

struct BiharmonicLeith{FT} <: AbstractScalarBiharmonicDiffusivity{HorizontalFormulation, 3}

The BiharmonicLeith struct represents a biharmonic Leith viscosity parameterization.

Fields

C::FT: The coefficient of the biharmonic Leith viscosity.

Reference

Fox-Kemper, B., & Menemenlis, D. (2004). Can large eddy simulation techniques improve mesoscale rich ocean models? In M. Hecht & H. Hasumi (Eds.), Ocean modeling in an eddying regime (pp. 319 - 337). doi: 10.1029/177GM19

BaroclinicAdjustment.Parameterizations.EnergyBackScatter — Type

struct EnergyBackScatter{FT} <: AbstractScalarDiffusivity{ExplicitTimeDiscretization, HorizontalFormulation, 2}

The EnergyBackScatter struct represents a parameterization for energy backscatter in a baroclinic adjustment model.

Fields

C₄: Coefficient C₄ for the biharmonic smagorinky dissipation

C₂: Coefficient C₂ for the backscattering antiviscosity

Cᴰ: Coefficient Cᴰ for the bottom drag dissipation

Uᵇ: Background velocity Uᵇ

τ: dissipation timescale τ

Reference

Jansen, M., Adcroft, A., Khani, S., & Kong, H. (2019). Toward an energetically consistent, resolution aware parameterization of ocean mesoscale eddies. Journal of Advances in Modeling Earth Systems, 11 (8), 2844-2860. doi: https://doi.org/10.1029/2019MS001750

BaroclinicAdjustment.Parameterizations.OMp25Closure — Type

struct OMp25Closure{FT} <: AbstractScalarDiffusivity{ExplicitTimeDiscretization, HorizontalFormulation, 2}

The OMp25Closure struct represents the lateral friction parameterization of the OMp25 GFDL ocean model.

Fields

C₄::FT: Coefficient C₄ for the bilaplacia Smagorinky viscosity

u₄::FT: Velocity scale u₄ for the static bilaplacian viscosity

C₂::FT: Coefficient C₂ for the laplacian Smagorinky viscosity

u₂::FT: Velocity scale u₂ for the static laplacian viscosity

Reference

Adcroft, A., Anderson, W., Balaji, V., Blanton, C., Bushuk, M., Dufour, C., . . . Zhang, R. (2019). The gfdl global ocean and sea ice model om4.0: Model description and simulation features. Journal of Advances in Modeling Earth Systems, 11 (10), 3167-3211. doi: https://doi.org/10.1029/2019MS001726633

BaroclinicAdjustment.Parameterizations.QGLeith — Type

struct QGLeith{FT, M, S} <: AbstractScalarDiffusivity{ExplicitTimeDiscretization, HorizontalFormulation, 2}

The QGLeith struct represents a parameterization for the Leith scalar diffusivity in a quasi-geostrophic model.

Fields

C: The coefficient of the Leith scalar diffusivity.

min_N²: The minimum value of the buoyancy frequency squared.

isopycnal_tensor: The isopycnal tensor used in the parameterization.

slope_limiter: The slope limiter used in the parameterization.

Reference

Bachman, S., Fox-Kemper, B., & Pearson, B. (2017). A scale-aware subgrid model for quasi-geostrophic turbulence. Journal of Geophysical Research: Oceans. doi: 10.1002/2016JC012265

BaroclinicAdjustment.Parameterizations.Δ²ᶜᶜᶜ — Method

Return the filter width for a Leith Diffusivity on a general grid.

BaroclinicAdjustment.Parameterizations.Δ̃ᶜᶜᶜ — Method

The averaged filter width